Description
So what I have done so far is that I created a GUI which shows 3 columns ("To Do", "Being Done" and "Done") and the the idea is that the user will be able to add a "Task" with the name, duration and importance of the task when they click a button and a rectangle with the name of the task would appear on the GUI. The status of the task always starts with "To Do". When a new task is added it is also added to an array list. Right now I haven't yet figured out a way the user could input the Task information in the GUI. Also I wanted to add a feature where the user would be able to drag the rectangle with the task name into the column with its current status ("To Do", "Being Done" and "Done").